整理id字段类型

+++ app/exec/backend/controllers/batchFrozenShortUrl.go
@@ -6,14 +6,14 @@ import (
 )
 
 type BatchFrozenShortUrlReq struct {
-	Ids      []int `from:"ids" valid:"Required"`
-	IsFrozen int   `from:"isFrozen" valid:"Range(0,1)"`
+	Ids      []uint32 `from:"ids" valid:"Required"`
+	IsFrozen uint8   `from:"isFrozen" valid:"Range(0,1)"`
 }
 
 type BatchFrozenShortUrlRes struct {
-	RequestCount int   `json:"requestCount"`
-	UpdateCount  int   `json:"updateCount"`
-	ErrIds       []int `json:"errIds"`
+	RequestCount uint32   `json:"requestCount"`
+	UpdateCount  uint32   `json:"updateCount"`
+	ErrIds       []uint32 `json:"errIds"`
 }
 
 // 函数名称: BatchFrozenShortUrl
@@ -41,17 +41,17 @@ func (c *BackendController) BatchFrozenShortUrl() {
 		return
 	}
 
-	var updateIds []int
-	errIds := make([]int, 0)
-	var insertShortNum []int
+	var updateIds []uint32
+	errIds := make([]uint32, 0)
+	var insertShortNum []uint32
 	// 提交id数量与查询出的数据量不一致
 	// 需要以数据库数据为准筛选出差集，准备进行错误返回
-	requestCount := len(req.Ids)
-	updateCount := len(data)
+	requestCount := uint32(len(req.Ids))
+	updateCount := uint32(len(data))
 	if updateCount != requestCount {
 
 		// 将请求操作的id 提为key
-		mapData := make(map[int]interface{})
+		mapData := make(map[uint32]interface{})
 		for _, v := range data {
 			mapData[v.Id] = v.ShortNum
 		}
@@ -59,7 +59,7 @@ func (c *BackendController) BatchFrozenShortUrl() {
 		for _, v := range req.Ids {
 			if mapData[v] != nil {
 				updateIds = append(updateIds, v)
-				insertShortNum = append(insertShortNum, mapData[v].(int))
+				insertShortNum = append(insertShortNum, mapData[v].(uint32))
 			} else {
 				errIds = append(errIds, v)
 			}
